speed_champion/speed_champion/api/circuits/serializers.py
from rest_framework import serializers
from .models import Circuit


class CircuitSerializer(serializers.ModelSerializer):
    class Meta:
        model = Circuit
        fields = ['id', 'name', 'city', 'type']
