pyproject.toml
[project]
name = "karts-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=6.0.1",
    "djangorestframework>=3.16.1",
    "mistralai>=1.2.4",
    "python-dotenv>=1.0.0",
    "pillow>=11.1.0",
    "requests>=2.32.3",
    "django-cors-headers>=4.9.0",
    "pillow-heif>=1.1.1",
]
